review: separeted  TestInstallSectionMerger Tests
https://github.com/skbkontur/cement/pull/69#discussion_r292302875
